Generate the Python third-party licenses page like the Node port
Adopt the Node docs mechanism (#11844): cargo-about's JSON output rendered
into a Starlight Markdown content page (src/content/docs/generated/
thirdparty.md), shown as a themed, searchable Third-Party Licenses entry at
the bottom of the sidebar, instead of an iframe to a standalone HTML asset.

Extract the shared generator into @slint/common-files/src/utils/thirdparty.ts
(renderThirdPartyMarkdown + generateThirdPartyMarkdown) and reduce both the
Node and Python generate-thirdparty.ts scripts to thin wrappers passing their
crate directory and output path. Drop the now-unused thirdparty.hbs and the
iframe page.
