+++ docs/python/.gitignore
@@ -1,10 +1,10 @@
-# cargo-about output (requires `cargo install cargo-about` in CI)
-public/thirdparty/
-
 # gen_mdx.py output (run via `pnpm gen`)
 src/content/docs/api/
 src/api-manifest.json
 src/version.json
+
+# cargo-about output (run via `pnpm thirdparty`)
+src/content/docs/generated/
 
 # build output
 dist/
