+++ docs/python/astro.config.mjs
@@ -54,9 +54,7 @@ export default defineConfig({
                 slintStarlightLinksValidatorPlugin({
                     exclude: ({ link }) => {
                         const p = (link.split("?")[0] ?? "").trim();
-                        // /thirdparty/ is a cargo-about asset (served under the
-                        // base path when deployed), not a content route.
-                        return p.startsWith("/#") || p.endsWith("/thirdparty/");
+                        return p.startsWith("/#");
                     },
                 }),
                 starlightExpandAllSidebarGroups(),
@@ -133,7 +131,7 @@ export default defineConfig({
                         },
                     ],
                 },
-                { label: "Third-Party Licenses", slug: "thirdparty-licenses" },
+                { autogenerate: { directory: "generated" } },
             ],
         }),
     ],
