+++ docs/python/scripts/generate-thirdparty.ts
@@ -1,22 +1,21 @@
 // Copyright © SixtyFPS GmbH <info@slint.dev>
 // SPDX-License-Identifier: MIT
-import { spawnSync } from "node:child_process";
-import { mkdirSync } from "node:fs";
 import { dirname, join } from "node:path";
 import { fileURLToPath } from "node:url";
+import { generateThirdPartyMarkdown } from "@slint/common-files/src/utils/thirdparty.ts";
 
 const scriptsDir = dirname(fileURLToPath(import.meta.url));
 const docsPythonRoot = join(scriptsDir, "..");
 const repoRoot = join(docsPythonRoot, "..", "..");
-const apiPython = join(repoRoot, "api", "python", "slint");
-const outFile = join(docsPythonRoot, "public", "thirdparty", "index.html");
 
-mkdirSync(dirname(outFile), { recursive: true });
-
-const result = spawnSync(
-    "cargo",
-    ["about", "generate", "thirdparty.hbs", "-o", outFile],
-    { cwd: apiPython, stdio: "inherit" },
-);
-
-process.exit(result.status ?? 1);
+generateThirdPartyMarkdown({
+    crateDir: join(repoRoot, "api", "python", "slint"),
+    outFile: join(
+        docsPythonRoot,
+        "src",
+        "content",
+        "docs",
+        "generated",
+        "thirdparty.md",
+    ),
+});
