--- docs/python/src/content/docs/thirdparty-licenses.mdx
@@ -1,15 +1,0 @@
----
-title: Third-Party Licenses
-description: Licenses of the dependencies bundled with the slint package.
----
-
-import { pythonDocsPublicAsset } from "../../python-site-config.mjs";
-
-The licenses below cover all third-party dependencies bundled with the `slint`
-package. They are generated from the crate dependency graph by `cargo about`.
-
-<iframe
-    src={pythonDocsPublicAsset("thirdparty/")}
-    title="Third-Party Licenses"
-    style="width: 100%; height: 80vh; border: 1px solid var(--sl-color-gray-5); border-radius: 0.5rem;"
-></iframe>
